Tidy up code, still occasionally wheels stop
Add comments into this code that is a lot experimented with.

Suspected problem is that 12 V voltage should be used for motor driver
instead of 8 V.

Hoping this issue will be fixed in the next commit, where there will
also be a clean code for car movement and light-adaptible LED working
fine.